third_party/blink/web_tests/fast/js/exception-thrown-from-eval-inside-closure-expected.txt

This page tests whether an exception thrown from an eval inside a closure prematurely tears off the closure's activation.

If the test passes, you'll see a PASS message below.

PASS: x should be 2 and is.
